+++ babel.js
@@ -1,6 +1,6 @@
 module.exports = {
     extends: [
         "eslint-config-konnektid",
-        "eslint-config-konnektid/rules/babel"
+        "eslint-config-konnektid/specs/es7"
     ]
 };