+++ es5.js
@@ -1,9 +1,0 @@
-module.exports = {
-    extends: [
-        "eslint-config-konnektid/rules/recommended",
-        "eslint-config-konnektid/rules/bestpractices",
-        "eslint-config-konnektid/rules/readability",
-        "eslint-config-konnektid/rules/snazzy",
-        "eslint-config-konnektid/rules/complexity"
-    ]
-};