+++ frontend.js
@@ -1,7 +1,11 @@
 module.exports = {
     extends: [
-        "eslint-config-konnektid/react",
-        "eslint-config-konnektid/rules/webpack",
-        "eslint-config-konnektid/rules/browser"
+        "eslint-config-konnektid/specs/es7",
+        "eslint-config-konnektid/specs/jsx",
+        "eslint-config-konnektid/envs/browser",
+        "eslint-config-konnektid/envs/webpack",
+        "eslint-config-konnektid/rules/base",
+        "eslint-config-konnektid/rules/frontend",
+        "eslint-config-konnektid/rules/react"
     ]
 };