+++ rules/development.js
@@ -6,6 +6,7 @@ module.exports = {
         "no-debugger": 0,
         "no-console": 0,
         "no-sync": 0,
+        "no-magic-numbers": 0,
 
         // relax rules
         "prefer-arrow-callback": 1,