+++ rules/frontend.js
@@ -1,8 +1,4 @@
 module.exports = {
-    env: {
-        node: false,
-        browser: true
-    },
     rules: {
 
         /*