+++ rules/react.js
@@ -2,11 +2,6 @@ module.exports = {
     plugins: [
         "react"
     ],
-    parserOptions: {
-        ecmaFeatures: {
-            jsx: true
-        }
-    },
     rules: {
 
         /*