+++ rules/test.js
@@ -1,12 +1,10 @@
 module.exports = {
-    env: {
-        mocha: true
-    },
     rules: {
         "global-require": 0,
         "no-unused-expressions": 0,
         "max-statements": 0,
         "require-jsdoc": 0,
+        "no-magic-numbers": 0,
 
         "no-new": 1,
         "new-cap": 1