+++ rules/universal.js
@@ -1,0 +1,7 @@
+module.exports = {
+    rules: {
+        
+        // allow conditional module inclusion
+        "global-require": 0
+    }
+};