+++ rules/webpack.js
@@ -1,8 +1,0 @@
-module.exports = {
-    globals: {
-
-        // webpack transforms node-style require calls
-        "require": false,
-        "module": false
-    }
-}