+++ specs/es5.js
@@ -1,0 +1,8 @@
+module.exports = {
+    env: {
+        es6: false
+    },
+    parserOptions: {
+        ecmaVersion: 5
+    }
+};