+++ specs/jsx.js
@@ -1,8 +1,7 @@
 module.exports = {
     parserOptions: {
-        ecmaVersion: 6,
         ecmaFeatures: {
-            experimentalObjectRestSpread: true
+            jsx: true
         }
     }
-}
+};