+++ specs/module.js
@@ -2,4 +2,4 @@ module.exports = {
     parserOptions: {
         sourceType: "module"
     }
-}
+};