+++ universal.js
@@ -1,0 +1,9 @@
+module.exports = {
+    extends: [
+        "eslint-config-konnektid/specs/es7",
+        "eslint-config-konnektid/envs/universal",
+        "eslint-config-konnektid/envs/webpack",
+        "eslint-config-konnektid/rules/base",
+        "eslint-config-konnektid/rules/universal"
+    ]
+};